imt/umt(4): Fix the match() and attach() functions.
The way imt(4) and umt(4) search for HID report IDs is broken;
hid_is_collection() is not suitable for that task.  Add a better helper
function to hid.c and adapt hidmt, imt, and umt to make use of it.

Thanks to Thomas Dettbarn for reporting the problem.

"makes (some) sense" to kettenis@