Separate a vertical line on the left side of a table by horizontal whitespace of 1n from the content of the leftmost column, both when the line results from a table option (e.g. "box") and when it results from a '|' layout option.
This is the same change as in
groff commit 8f06678 Feb 3 02:22:02 2023 -0600 by gbranden@,
which was released with groff-1.23.0.

This does change the appearance of existing documents that use tbl(7),
so technically, it's a backward-incompatible change, but it is highly
unlikely to degrade formatting in any situation, it almost always looks
better, and anyway, compatibility with modern groff matters more than
compatibility with historical formatters and documents.