Like for .Dd, the code changes to .Dt in groff-1.23 are so extensive that with reasonable effort, only part of the regressions can be patched back. If i went down that rabbit hole now, i would never get groff updated, so the best i can do for now is to mark the affected tests with SKIP_GROFF.
Essentially, groff-1.23 is completely broken unless authors stick to the
conventional order of the prologue macros, i.e. .Dd .Dt .Os.  OpenBSD base
system manual pages are not affected because we always stick to that order,
but it would be somewhat surprising if there were no third-party pages
somewhere in the wild using a different order.  We'll likely have to
return to this particular regression at a later time.