Upgrade vmd(8)'s virtio implementation to support v1.x.
This updates the device emulation for the block, entropy, network,
and scsi virtio devices to use virtio 1.x.

Not only does this simplify a device emulation, it allows us to
start adding newer virtio device types and take advantage of
newer networking features (once implemented).

This is a hard upgrade with no support for legacy modes, so guests
that do not support virtio 1.0 or newer will be unhappy. In OpenBSD,
that means pre-6.6. For Linux, this means kernels v4.0 and newer
circa 2015.

Lots of testing from bluhm@, hshoexer, mlarkin@. Feedback from sf@.

go for it, @mlarkin