Disable "status memory" command.
About a year ago we had to disable the shared cache for the different
resolver strategies because it led to a segfault in unwind after a
libunbound update. There are no more memory statistics to show at this
point. I am still hopeful to bring the feature back one day, so hide
the command behind a ifdef like we did in resolver.c.

Missing feature reported by Jordan Geoghegan on bugs@