Remove net.inet6.ip6.soiikey sysctl
When we implemented RFC 7217 - "A Method for Generating Semantically
Opaque Interface Identifiers with IPv6 Stateless Address
Autoconfiguration (SLAAC)" we added support for global addresses as
well as link local addresses.

We quickly learned that some hosting providers insisted on link-local
addresses derived from mac-addresses (eui64). They would only route
IPv6 traffic to those. So we disabled RFC 7217 for link-local
addresses.

This means that we no longer need the sysctl, slaacd(8) can just read
/etc/soii.key directly instead of bothering the kernel.

These are the userland bits, I'll commit the kernel part in a week or
so, not everybody keeps kernel and userland in sync.

Input & OK bluhm