Document reachable time and retrans timer settings.
IPv6 has many buttons to push and knobs to twiddle...

These two are in the router advertisement packet header, so we have to
send something (0). slaacd ignores the values and there is no way to
set the values in the kernel, either on OpenBSD. No idea what other
systems do with this.

Prodded by schwarze some time ago.