Remove upper layer neighbor reachability hints.
Running neighbor discovery once in a while is better than touching the
neighbor cache for (nearly) every TCP packet.

This functionality was disabled by default, however due to a bug /
missing optimization, the code did a bunch of work before figuring out
that the feature is disabled.

With this, net.inet6.icmp6.nd6_maxnudhint can be removed.

OK bluhm