Abort transfers from servers that send excessive data.
When a regular HTTP transfer with content-length is finished check that
the buffer is actually empty. It seems some proxies violate this and
then rpki-client runs into an assert in http_done(). Instead fail the
transfer and let the state machine figure out how to fall back.

Reported by and OK job@