reduce default so-sndbuf in unbound to avoid a warning.
in 1.24.0, upstream started to override the OS default and warn at
startup if not possible, but they picked a value (4MB) which is above
the 2MB maximum that we permit.

found by landry@, ok landry bluhm

from bluhm:

"Our UDP stack has no UDP socket buffer for sending.  Each packet
is directly passed to ip_output().  So basically this value is a
per packet size limit."