implement IFCAP_TSO support.
this allows the stack to pass large tcp frames between rdomains.

note that rport doesn't advertise LRO support, which mostly means
that LRO is not something that can be disabled on rport interfaces.
as an l3 tunnel interface, there's no topology that an rport interface
can be set up in where disabling LRO is needed, unlike l2 Ethernet
interfaces. coordinating disabling LRO on an rport and then clearing
the TSO flags on the partner rport interface is not worth the effort
to actually implement the capability isn't worth it.