update the match code for quectel devices.
this is based on information in the Quectel UMTS LTE 5G Linux USB
Driver User Guide V3.3 documentation, which is the latest version
i can find.

it works on EC200A devices, which benefit from this change, and on
EC25s, which were working before this change and still work after.