The pmap_table array comes from Mach on luna88k and is used to specify what memory areas should be mapped in the kernel.
In Mach, this table contained the main memory, which is why it had a
cacheability field. In BSD, main memory is handled separately, and the
pmap_table is only used to map the few hardware resources for which a driver
does not necessarily attaches - or which needs to be used by the kernel before
a driver gets a chance to attach.

Drop the cacheability field, since all entries will require cache inhibited
mappings.

Tested & ok aoyama@