libsndio: Don't use the hdl->wused counter to trigger AUDIO_START
No behavior change.

The hdl->wused variable counts the amount of data the program writes.
It doesn't take into account the silence inserted with sio_wsil().
AUDIO_START must be called when a full buffer is written to the device
regardless whether its silence or data. In the current implementation,
we don't insert silence (yet), so this commit is mostly to avoid
mistakes in future changes.