add ioctls for managing a vlan aware bridge.
SIOCBRDGSVADDR can be used to add an address to the forwarding
database, like SIOCBRDGSADDR, but the vlan the mac address is scoped
to can be specified.

SIOCBRDGDVADDR can be used to remove an address from the forwarding
database, like SIOCBRDGDADDR, but the vlan the mac address is scoped
to can be specified.

SIOCBRDGSPVID allows the pvid (native vlan or the vlan untagged
traffic belongs to) on a port to be configured.

SIOCBRDGSVMAP is used to manipulate the set of vids (vlan ids) that
a port will handle tagged traffic for. SIOCBRDGGVMAP can be used
to get a copy of the vid map.

SIOCBRDGVRTS is an extended version of SIOCBRDGRTS for dumping
bridge forwarding entries with vlan information for each entry.

if.c enforces that only root can use SIOCBRDGSVADDR, SIOCBRDGDVADDR,
SIOCBRDGSPVID, and SIOCBRDGSVMAP.