add support for managing a vlan aware bridge, ie, veb(4).
the "static" and "deladdr" commands have been extended to recognise
a macaddr@vid syntax. if a vid is specified it'll try and use the
SIOCBRDGSVADDR and SIOCBRDGDVADDR ioctls instead of SIOCBRDGSADDR
and SIOCBRDGDADDR respectively.

when printing information about a bridge, the code will try to using
SIOCBRDGVRTS to dump address entries with vid information. if that's
not supported it will fall back to the existing SIOCBRDGRTS code.
when printing info about bridge ports it will also try and use
SIOCBRDGSVMAP to fetch the vid map.

this adds an "untagged" command for configuring the pvid on a veb
port. "-untagged" can be used to disable the pvid on a port.

"tagged" is used to configure the vidmap on a port. it currently
only supports handling one vid at a time. if the vid is prefixed
with a + it will ask the kernel to add the vid to the vidmap, and
a - prefix will remove that vid from the map. a vid without a prefix
will replace the map.